refactor: 리스트 컴포넌트 구조 개선 및 통합
  - ListComponent 신규 생성으로 재사용 가능한 리스트 아이템 컴포넌트 구현
  - ListLayout으로 배열 데이터 렌더링 로직 통합
  - Test_* 컴포넌트들 제거 및 기능 통합
  - TogetherList, Test_ReviewList 컴포넌트 구조 개선
  - /test 페이지에서 새로운 컴포넌트 패턴 테스트 환경 구축
