+++ app/test/page.jsx
@@ -1,8 +1,8 @@
 import TestReviewList from "@/components/community/Test_ReviewList";
-import ListGallery from "@/components/global/ListView";
-import List from "@/components/global/Test_List";
-import TestListLayout from "@/components/global/Test_ListLayout";
-import TestTogetherList from "@/components/together/Test_TogetherList";
+import ListGallery from "@/components/global/ListComponent";
+import List from "@/components/global/ListComponent";
+import TestListLayout from "@/components/global/ListLayout";
+import TestTogetherList from "@/components/together/TogetherList";
 import TogetherList from "@/components/together/TogetherList";
 
 export default function test() {
