--- components/global/Test_ListLayout.jsx
@@ -1,9 +1,0 @@
-export default function TestListLayout({ Component, items, commonProps={} }) {
-  return (
-    <div className="flex flex-col gap-0 w-full">
-      {items.map((item, idx) => (
-        <Component key={idx} {...item} {...commonProps} />
-      ))}
-    </div>
-  )
-}
