+++ components/together/TogetherList.jsx
@@ -1,43 +1,50 @@
-import ListGallery from "../global/ListView";
 import { ICONS } from "@/constants/path";
 import Image from "next/image";
+import ListComponent from "../global/ListComponent";
 
 export default function TogetherList({
   imgSrc,
   title = "모집글 제목",
   eventType = "이벤트유형",
   eventName = "이벤트명",
-  group = "00/00",
+  people = "00/00",
   date = "0000.00.00",
-  pin = "00시 00구 00동",
-  nickName = "사용자 별명",
-  alt,
+  address = "00시 00구 00동",
+  author = "-",
+  alt = "",
+  isClosed = false,
 }) {
   return (
-    <>
-      <div className="flex flex-wrap w-full">
-        <ListGallery src={imgSrc} alt={alt} />
-        <div className="mt-6">
-          <div className="flex items-center space-x-2 text-gray-400 opacity-70">
-            <div className="border border-b-2 rounded-4xl px-2">
-              {eventType}
-            </div>
-            <div>{eventName}</div>
+    <div className="relative">
+      {isClosed && <div className="absolute inset-0 w-full h-full bg-black opacity-10 z-10" />}
+      <ListComponent src={imgSrc} alt={alt}>
+        <div className="flex flex-col justify-around h-full">
+          <div className="flex gap-2">
+            <span className="border border-b-2 rounded-4xl px-2 w-fit">{eventType}</span>
+            <strong>{eventName}</strong>
           </div>
-          <div className="text-base font-bold">{title}</div>
-          <div className="flex items-center space-x-2 text-gray-400 opacity-70">
-            <Image src={ICONS.CALENDAR} alt="calendar" width={24} height={24} />
-            <div>{date}</div>
-            <Image src={ICONS.GROUP} alt="group" width={24} height={24} />
-            <div>{group}</div>
-            <Image src={ICONS.PIN} alt="pin" width={24} height={24} />
-            <div>{pin}</div>
+          <h3 className="text-lg font-bold overflow-hidden whitespace-nowrap text-ellipsis text-black">
+            {title}
+          </h3>
+          <div className="flex gap-4 shrink-0 w-full">
+            <span className="flex items-center gap-2 flex-shrink-0">
+              <Image src={ICONS.CALENDAR} alt="calendar" width={20} height={20} />
+              {date}
+            </span>
+            <span className="flex items-center gap-2 flex-shrink-0">
+              <Image src={ICONS.GROUP} alt="group" width={24} height={24} />
+              {people}
+            </span>
+            <span className="flex items-center gap-2 flex-shrink-0">
+              <Image src={ICONS.PIN} alt="group" width={20} height={20} />
+              {address}
+            </span>
           </div>
-          <div className="space-x-2 text-gray-400 opacity-70">
-            작성자 : {nickName}
+          <div className="">
+            작성자 : {author}
           </div>
         </div>
-      </div>
-    </>
+      </ListComponent>
+    </div>
   );
 }
